backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
pydantic>=2.0
python-dotenv
httpx
python-multipart
